[add] 예측치 저장을 위한 코드

+++ src/main/java/energypa/bems/predictElec/dto/RequestElecDto.java
@@ -12,7 +12,7 @@ public class RequestElecDto {
     private String timestamp;
     private Integer Building;
     private Integer Floor;
-    private Integer Consumption;
+    private Double Consumption;
 
 
 }
